Move integration test setup to ethereumtest
The integration test setup in client package for initialzing a chain
backend with deployed contracts is chain specific functionality.

Also from ethereumtest package, this can be used in other integration
tests that will be implemented later.

Resolves #78
